Add reader capable of providing input to process via argc
This relies on SYS V process initialization, which puts the argc count in
the first stack.
This run-reader will pass enough arguments to give argc the specified value,
which is sufficient to provide input to the program now that the stack uses the
SYS V stack.

Undocumented, no examples yet.