Implement fvars properly in interpreters
fvars are now compiled to stack locations, and frames are properly implemented
around non-tails.

Because their implementation uses let-syntax and make-set!-transformers, having
a large number of fvars statically generated (as we did previously) causes
performance problems.
As a workaround, I've lowed the number, but this could cause issues with large
tests.
A proper fix would required a (quick) traversal of the input program to identify
the largest fvar, and bind up to that one.